chore(document): remove function injection in PDF converter
